Add disassembler to the library
Functions to perform basic disassembling have been added to the library.

Because the program may contain arbitrary data, some recovery from
unknown values in the input is required. Currently, the recovery is
extremely basic: if the next int is not recognised as an instruction, it
is disassembled as data and the process moves on to the next int in the
stream. This also means that data that happens to be a valid instruction
may cause the disassemble to fail to recognise actual instructions later
in the stream.